Add files via upload
"Write a C# program that implements an abstract class Bird that has a Name property of type text and three methods SetName (string name), GetName and Fly. 
The Eat method will be an abstract method of type void.
You will also need to create a Crow/Parrot class that implements the above Bird class and the Fly method should say that the crow/parrot is flying. 
To test the program ask the user for a bird name and create a new Crow/Parrot type object from the Main program, 
give the Crow/Parrot object a name, and then execute the GetName and Fly methods."