Add files via upload
"Write a C# program that uses a Queue to store a list of people. Create a class called Person.cs with two properties (name and age) 
and a ToString() method to print the result.

Then define a Queue and request the information (name and age) of 5 people from the user, you will have to 
store the information in the queue.  Finally print the result of the queue on screen. "