Add files via upload
"Create a C# program to:
The main task is to get a stage / event ready with the following tasks:
Flower decorations to be completed on stage.
Chief guest to be picked up from the airport. 
Co-ordinate with the caterers and set up the food area. 
Arrange the gifts to be distributed in the stage after the decorations are over. 
Start the function only when the Chief guest has arrived at the venue. "