Add files via upload
"Write a C# program to create a file with a name given by the user. Add atleast 10 lines of content to the file. 
Get 2 inputs from the user, old word and new word and replace the old word with the new word throughout the file. 
You can show a little bit of content to the user as hints for the old word. "