Add files via upload
"Write a C# program to create a Phonebook. 
The Phonebook should take inputs from the user and store their numbers. 
Also ask for the user age and store their numbers as Students in a sub class. If the user gives a profession, store the numbers in the Service subclass. 
Display the numbers as a collection.